REFC: change file structure for test
